veltrix-backend/requirements.txt
fastapi==0.121.1
uvicorn[standard]==0.29.0
pydantic==2.12.3
pydantic-settings==2.11.0
scikit-learn==1.7.2
numpy==2.2.6
python-dotenv==1.0.1
httpx==0.27.0
beautifulsoup4==4.12.3
tldextract==5.1.2
joblib==1.4.2
python-multipart==0.0.9
